refactor: split stack & queue
